Fixed bug reading HplProfile. Seconds should have been double not int when calculated from decimal hour and the way they were cast as ints was wrong anyway, making the always 0.

+++ LidarQuicklookPlotter/HplProfile.cpp
@@ -22,7 +22,7 @@ bool HplProfile::readFromStream(std::istream &stream, const HplHeader &header)
 
 	unsigned int hour = (unsigned int)std::floor(decimalHour);
 	unsigned int minute = (int)std::floor((decimalHour - hour)*60.0);
-	unsigned int second = (int)((decimalHour - hour)*60.0 - minute) * 60;
+	double second = ((decimalHour - hour)*60.0 - minute) * 60;
 	m_time = header.startTime;
 	m_time.setTime(hour, minute, second);
